--- app/src/main/java/es/architectcoders/spaceexplorer/components/home/HomeFragment.kt
@@ -1,29 +1,0 @@
-package es.architectcoders.spaceexplorer.components.home
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import dagger.hilt.android.AndroidEntryPoint
-import es.architectcoders.spaceexplorer.R
-
-@AndroidEntryPoint
-class HomeFragment : Fragment() {
-
-    private val viewModel: HomeViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_home, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        viewModel.init()
-    }
-
-}
