+++ app/src/main/java/es/architectcoders/spaceexplorer/ui/favourite/FavouriteViewModel.kt
@@ -1,10 +1,10 @@
-package es.architectcoders.spaceexplorer.components.rovers
+package es.architectcoders.spaceexplorer.ui.favourite
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class RoversViewModel @Inject constructor() : ViewModel() {
+class FavouriteViewModel @Inject constructor() : ViewModel() {
     // TODO: Implement the ViewModel
 }
