+++ app/src/main/java/es/architectcoders/spaceexplorer/ui/mars/MarsViewModel.kt
@@ -1,4 +1,4 @@
-package es.architectcoders.spaceexplorer.components.mars
+package es.architectcoders.spaceexplorer.ui.mars
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
