+++ app/src/main/java/es/architectcoders/spaceexplorer/ui/rovers/RoversViewModel.kt
@@ -1,10 +1,10 @@
-package es.architectcoders.spaceexplorer.components.favourite
+package es.architectcoders.spaceexplorer.ui.rovers
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class FavouriteViewModel @Inject constructor() : ViewModel() {
+class RoversViewModel @Inject constructor() : ViewModel() {
     // TODO: Implement the ViewModel
 }
